feat: edit and save mode on hole section not single property
